% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableTests.R
\name{availableTests}
\alias{availableTests}
\title{availableTests}
\usage{
availableTests()
}
\value{
a character vector of available tests that can be used as input in \code{\link{checkArgus}}
}
\description{
checks which test-scenarios are defined. The keywords returned here can be used as inputs to \code{\link{checkArgus}}.
}
\examples{
\dontrun{
avail <- availableTests()
checkArgus(tau_exe="c:/tau/tau.exe", what=avail[1])
}
}
