% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUITests.R
\name{GUITests}
\alias{GUITests}
\title{GUITests}
\usage{
GUITests()
}
\value{
a browser containing the document will be opened.
}
\description{
shows a document specifiying tests for the graphical user interface of tau-argus.
}
\examples{
\dontrun{
GUITests()
}
}
